Move QP lifecycle tests to a dedicated module
